+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/OutboundProduct.java
@@ -95,6 +95,6 @@ public class OutboundProduct {
     private Long linenum;
 
     @ApiModelProperty("税率")
-    private Long tax;
+    private BigDecimal tax;
 
 }
